refactor: handler を機能別サブパッケージに分割
- handler/{content,user,favorite,other}/ に操作単位でファイルを分割
- 各パッケージに handler.go（構造体定義）+ 操作ごとのファイル構成
- infra/routing.go のimportをサブパッケージエイリアスに更新
- ロジック変更なし（go build / go vet エラーなし）
