+++ t0016Go/handler/content/delete_vtuber.go
@@ -1,0 +1,52 @@
+package content
+
+import (
+	"log"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sharin-sushi/0016go_next_relation/common"
+	"github.com/sharin-sushi/0016go_next_relation/domain"
+)
+
+func (h *ContentHandler) DeleteVtuber(c *gin.Context) {
+	listenerId, err := common.TakeListenerIdFromJWT(c)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Error fetching listener info",
+		})
+		return
+	}
+	var selectedVtuber domain.Vtuber
+	if err := c.ShouldBind(&selectedVtuber); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invalid request body",
+		})
+		return
+	}
+
+	if isAuth, err := h.ContentService.VerifyUserModifyVtuber(listenerId, selectedVtuber); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Auth Check is failed.(we could not Verify)",
+		})
+		return
+	} else if !isAuth {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Only The Inputter can modify each data",
+		})
+		return
+	}
+
+	if err := h.ContentService.DeleteVtuber(selectedVtuber); err != nil {
+		log.Println("err:", err)
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Only Inputter can modify each data",
+			"error":   err,
+		})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Successfully Delete",
+	})
+}
