+++ t0016Go/handler/content/edit_karaoke.go
@@ -1,0 +1,54 @@
+package content
+
+import (
+	"log"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sharin-sushi/0016go_next_relation/common"
+	"github.com/sharin-sushi/0016go_next_relation/domain"
+)
+
+func (h *ContentHandler) EditKaraoke(c *gin.Context) {
+	listenerId, err := common.TakeListenerIdFromJWT(c)
+	if err != nil {
+		log.Println("err:", err)
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Error fetching listener info",
+		})
+		return
+	}
+
+	var Karaoke domain.Karaoke
+	if err := c.ShouldBind(&Karaoke); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invalid request body",
+		})
+		return
+	}
+
+	Karaoke.KaraokeInputterId = listenerId
+	if isAuth, err := h.ContentService.VerifyUserModifyKaraoke(listenerId, Karaoke); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Auth Check is failed.(we could not Verify)",
+		})
+		return
+	} else if !isAuth {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Only The Inputter can modify each data",
+		})
+		return
+	} else {
+		log.Printf("isAuth%v :\n", isAuth)
+		if err := h.ContentService.UpdateKaraoke(Karaoke); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{
+				"message": "Inputter can modify each data",
+			})
+			return
+		}
+	}
+
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Successfully Update",
+	})
+}
