+++ t0016Go/handler/content/edit_movie.go
@@ -1,0 +1,54 @@
+package content
+
+import (
+	"log"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sharin-sushi/0016go_next_relation/common"
+	"github.com/sharin-sushi/0016go_next_relation/domain"
+)
+
+func (h *ContentHandler) EditMovie(c *gin.Context) {
+	listenerId, err := common.TakeListenerIdFromJWT(c)
+	if err != nil {
+		log.Println("err:", err)
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Error fetching listener info",
+		})
+		return
+	}
+
+	var Movie domain.Movie
+	if err := c.ShouldBind(&Movie); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invalid request body",
+		})
+		return
+	}
+
+	Movie.MovieInputterId = listenerId
+	if isAuth, err := h.ContentService.VerifyUserModifyMovie(listenerId, Movie); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Auth Check is failed.(we could not Verify)",
+		})
+		return
+	} else if !isAuth {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Only The Inputter can modify each data",
+		})
+		return
+	}
+
+	if err := h.ContentService.UpdateMovie(Movie); err != nil {
+		log.Println("err:", err)
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Inputter can modify each data",
+		})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Successfully Update",
+	})
+}
