+++ t0016Go/handler/favorite/delete_karaoke.go
@@ -1,0 +1,41 @@
+package favorite
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sharin-sushi/0016go_next_relation/common"
+	"github.com/sharin-sushi/0016go_next_relation/domain"
+)
+
+func (h *FavoriteHandler) DeleteKaraokeFavorite(c *gin.Context) {
+	applicantListenerId, err := common.TakeListenerIdFromJWT(c)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Error fetching listener info",
+		})
+		return
+	}
+	var fav domain.Favorite
+	if err := c.ShouldBind(&fav); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invalid request body",
+		})
+		return
+	} else if fav.MovieUrl == "" || fav.KaraokeId == 0 {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Send Data has \"\" or 0",
+		})
+		return
+	}
+	fav.ListenerId = applicantListenerId
+	if err := h.ActivityService.DeleteKaraokeFavorite(fav); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invailed UnFavorite it",
+		})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Successfully UnFavorite it",
+	})
+}
