+++ t0016Go/handler/other/handler.go
@@ -1,4 +1,4 @@
-package handler
+package other
 
 import (
 	"net/http"
