--- t0016Go/handler/update_data.go
@@ -1,5 +1,0 @@
-package handler
-
-// 手動でmerge, buildしてる限りは、apiをbuildし直す日時をメモする（たまに忘れる）
-// 曲名１文字対応、等
-const LastUpdateData = "2024/01/14 3時 fix:歌複数登録要のapiを追加"
