+++ t0016Go/handler/user/profile.go
@@ -1,0 +1,35 @@
+package user
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sharin-sushi/0016go_next_relation/common"
+)
+
+func (h *UserHandler) GetListenerProfile(c *gin.Context) {
+	ListenerId, err := common.TakeListenerIdFromJWT(c)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Need Login"})
+		return
+	}
+
+	ListenerInfo, err := h.UserService.FindUserByListenerId(ListenerId)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error fetching listener info"})
+		return
+	}
+
+	fmt.Printf("ListenerInfo = %v \n", ListenerInfo)
+
+	c.JSON(http.StatusOK, gin.H{
+		"ListenerId":   ListenerInfo.ListenerId,
+		"ListenerName": ListenerInfo.ListenerName,
+		"CreatedAt":    ListenerInfo.CreatedAt,
+		"UpdatedAt":    ListenerInfo.UpdatedAt,
+		"Email":        "secret",
+		"Password":     "secret",
+		"message":      "got urself infomation",
+	})
+}
