+++ t0016Go/handler/user/withdraw.go
@@ -1,0 +1,43 @@
+package user
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sharin-sushi/0016go_next_relation/common"
+	"github.com/sharin-sushi/0016go_next_relation/domain"
+)
+
+func (h *UserHandler) LogicalDeleteUser(c *gin.Context) {
+	tokenLId, err := common.TakeListenerIdFromJWT(c)
+	fmt.Printf("tokenLId = %v \n", tokenLId)
+
+	if err != nil {
+		fmt.Printf("err:LogicalDeleteUser TakeListenerIdFromJWT, %v\n", err.Error())
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invalid ListenerId of token",
+			"err":     err,
+		})
+		return
+	} else if tokenLId == guestID {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Guest Acc. must NOT Withdrawal",
+		})
+		return
+	}
+	var dummyLi domain.Listener
+	dummyLi.ListenerId = tokenLId
+	if err := h.UserService.LogicalDeleteUser(dummyLi); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"message": "Invalid Withdrawn",
+			"err":     err,
+		})
+		return
+	}
+
+	c.SetCookie("auth-token", "none", -1, "/", "localhost", false, true)
+	c.JSON(http.StatusOK, gin.H{
+		"message": "Successfully Withdrawn. You can restore ur acc. within 60 days.",
+	})
+}
